+++ vendas.h
@@ -1,11 +1,11 @@
 typedef struct venda{
     int codigo;
     int quantidade;
-    double montante;
+    int montante;
 } Venda;
 
 //API
-Venda newVenda(int codigo, int quantidade, double montante);
+Venda newVenda(int codigo, int quantidade, int montante);
 int getVenda(off_t code, Venda *v);
 void saveVenda(Venda v);
 void translateVendas();
